requirements.txt
PyYAML
openai
opencc-python-reimplemented
numpy # "numpy<2" for faiss-gpu compatibility
faiss # faiss-gpu
jieba
rank_bm25